[ExecuTorch] Refactor fast_hadamard_transform_test shared implementation functions
Following diff will use this stuff to test the ExecuTorch operator as well.

Differential Revision: [D62760050](https://our.internmc.facebook.com/intern/diff/D62760050/)

[ghstack-poisoned]
