+++ extension/llm/custom_ops/spinquant/test/targets.bzl
@@ -6,12 +6,26 @@ def define_common_targets():
     The directory containing this targets.bzl file should also contain both
     TARGETS and BUCK files that call this function.
     """
+    runtime.cxx_library(
+        name = "fast_hadamard_transform_test_impl",
+        srcs = ["fast_hadamard_transform_test_impl.cpp"],
+        headers = [
+            "fast_hadamard_transform_special_unstrided_cpu.h",
+            "fast_hadamard_transform_test_impl.h",
+        ],
+        exported_deps = [
+            "//executorch/extension/llm/custom_ops/spinquant:fast_hadamard_transform",
+        ],
+        deps = [
+            "//executorch/extension/llm/custom_ops/spinquant/third-party/FFHT:dumb_fht",
+        ],
+    )
+
     runtime.cxx_test(
         name = "fast_hadamard_transform_test",
         srcs = ["fast_hadamard_transform_test.cpp"],
-        headers = ["fast_hadamard_transform_special_unstrided_cpu.h"],
         deps = [
+            ":fast_hadamard_transform_test_impl",
             "//executorch/extension/llm/custom_ops/spinquant:fast_hadamard_transform",
-            "//executorch/extension/llm/custom_ops/spinquant/third-party/FFHT:dumb_fht",
         ],
     )
